float to double to float

+++ inc/intersection/i_utils.h
@@ -18,15 +18,15 @@
 
 typedef struct s_t4
 {
-	double	t1;
-	double	t2;
-	double	t3;
-	double	t4;
+	float	t1;
+	float	t2;
+	float	t3;
+	float	t4;
 }	t_t4;
 
 t_t4	check_t_values(t_t4 t);
 t_t4	quadratic_formula(t_xyz input1, t_xyz input2,
-			double input3, double input4);
+			float input3, float input4);
 float	get_intersect_with_cap_planes(t_ray ray, t_object cylinder, float flag);
 
 #endif
