+++ inc/utils/u_vector_math.h
@@ -16,10 +16,10 @@
 # include "main.h"
 # include "utils/u_vector_operators.h"
 
-double	get_vector_length(t_xyz vector);
-double	get_dot_product(t_xyz v1, t_xyz v2);
+float	get_vector_length(t_xyz vector);
+float	get_dot_product(t_xyz v1, t_xyz v2);
 void	normalize_vector(t_xyz *vector);
-double	get_angle_between_vectors(t_xyz v1, t_xyz v2);
+float	get_angle_between_vectors(t_xyz v1, t_xyz v2);
 t_xyz	get_cross_product(t_xyz v1, t_xyz v2);
 
 #endif 		// VECTOR_MATH
